+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-* Added "None" option in modifier settings (fixed).
-* Added "Summons 2" option to duplication behavior when modifier is hold.
+* Added grid display for mounts.